+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 0.1.3
 
 * Added functions: `to_degrees`, `to_radians`, `hi`, `lo`.
+* `is_valid` method checks for overlapping representations.
 
 ## Version 0.1.2